fix bug with season switching not working
ended up being a race condition where the server told the client to write
the user cookie before the client had setup the event to write the actual
cookie. Then when the browser would refresh it would give that socket a
new user hash and therefor the store wouldn't have the season id set so it
would constantly go to the current season. Ended up overriding the socket
emit event on the server to queue any calls to it. Then once the client
was ready it emits a ready event which replaces the emit function and
handles any queued events